chromeos-dbus-bindings: Generator must expose property name accessors
D-Bus property names tend to be used in client code literally to
validate the property name in property change notifications, etc.

This forces clients to hard-code the name in their code and when
an interface definition changes in XML and new adaptor/proxies are
generated, the client code is not updated but still compiles, so the
breaking change is hard to notice.

Expose property name accessors in the generated code which makes the
client break at compile time if the property name is changed.

BUG=brillo:348
TEST=`FEATURES=test emerge-link chromeos-dbus-bindings`

Change-Id: I21bb2c06faf134aa92c923085ddeebce16806fb8
Reviewed-on: https://chromium-review.googlesource.com/252420
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>